% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrangeDS.R
\name{arrangeDS}
\alias{arrangeDS}
\title{Order the rows of a data frame by the values of selected columns}
\usage{
arrangeDS(expr, .data, .by_group)
}
\arguments{
\item{expr}{Variables, or functions of variables. Use desc() to sort a variable in descending
order.}

\item{.data}{A data frame, data frame extension (e.g. a tibble), or a lazy data frame (e.g. from dbplyr or dtplyr).}

\item{.by_group}{If TRUE, will sort first by grouping variable. Applies to grouped data frames
only.}
}
\value{
the object specified by the \code{newobj} argument of \code{ds.arrange} which is written
to the serverside.
}
\description{
DataSHIELD implentation of  \code{dplyr::arrange}.
}
