% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bindColsDS.R
\name{bindColsDS}
\alias{bindColsDS}
\title{Bind multiple data frames by column}
\usage{
bindColsDS(to_combine = NULL, .name_repair = NULL)
}
\arguments{
\item{to_combine}{Data frames to combine. Each argument can either be a data frame, a list that
could be a data frame, or a list of data frames. Columns are matched by name, and any missing
columns will be filled with NA.}

\item{.name_repair}{One of "unique", "universal", or "check_unique". See
\code{vctrs::vec_as_names()} for the meaning of these options.}
}
\value{
A serverside data frame with name specified in \code{newobj} and the same type as the
first element of `to_combine`.
}
\description{
DataSHIELD implementation of \code{dplyr::bind_cols}.
}
