% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind_rowsDS.R
\name{bindRowsDS}
\alias{bindRowsDS}
\title{Bind multiple data frames by row.}
\usage{
bindRowsDS(to_combine = NULL, .id = NULL)
}
\arguments{
\item{to_combine}{Data frames to combine. Each argument can either be a data frame, a list that
could be a data frame, or a list of data frames. Columns are matched by name, and any missing
columns will be filled with NA.}

\item{.id}{he name of an optional identifier column. Provide a string to create an output column
that identifies each input. The column will use names if available, otherwise it will use
positions.}
}
\value{
A serverside data frame with name specified in \code{newobj} and the same type as the
first element of `to_combine`.
}
\description{
DataSHIELD implementation of \code{dplyr::bind_rows}.
}
