% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case_whenDS.R
\name{case_whenDS}
\alias{case_whenDS}
\title{Performs dplyr if_else}
\usage{
case_whenDS(dynamic_dots = NULL, default = NULL, .ptype = NULL, .size = NULL)
}
\arguments{
\item{dynamic_dots}{A sequence of two-sided formulas. The left hand side (LHS) determines which
values match this case. The right hand side (RHS) provides the replacement value.
The LHS inputs must evaluate to logical vectors.
The RHS inputs will be coerced to their common type.
All inputs will be recycled to their common size. That said, we encourage all LHS inputs to be
the same size. Recycling is mainly useful for RHS inputs, where you might supply a size 1 input
that will be recycled to the size of the LHS inputs.
NULL inputs are ignored.}

\item{default}{The value used when all of the LHS inputs return either FALSE or NA.}

\item{.ptype}{An optional prototype declaring the desired output type. If supplied, this overrides the common type of true, false, and missing.}

\item{.size}{An optional size declaring the desired output size. If supplied, this overrides the size of condition.}
}
\value{
the object specified by the \code{newobj} argument
of \code{ds.rename} (or as default same name as input object) which is written to the serverside.
}
\description{
This function is similar to R function \code{rename}.
}
\details{
Performs dplyr rename
}
