% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyFillDS.R
\name{classAllColsDS}
\alias{classAllColsDS}
\title{Get the Class of All Columns in a Data Frame}
\usage{
classAllColsDS(df.name)
}
\arguments{
\item{df.name}{A string representing the name of the data frame.}
}
\value{
A tibble with the class of each column in the data frame.
}
\description{
Get the Class of All Columns in a Data Frame
}
