% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyFillDS.R
\name{convert_class}
\alias{convert_class}
\title{Convert a Vector to a Specified Class}
\usage{
convert_class(x, class_name)
}
\arguments{
\item{x}{The vector to be converted.}

\item{class_name}{A string indicating the target class (1 = factor, 2 = integer, 3 = numeric,
4 = character, 5 = logical).}
}
\value{
The converted vector.
}
\description{
Convert a Vector to a Specified Class
}
