% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distinctDS.R
\name{distinctDS}
\alias{distinctDS}
\title{Keep distinct/unique rows}
\usage{
distinctDS(expr, .data, .keep_all)
}
\arguments{
\item{expr}{Optional variables to use when determining uniqueness. If there are multiple rows for
a given combination of inputs, only the first row will be preserved. If omitted, will use all
variables in the data frame.}

\item{.data}{A data frame, data frame extension (e.g. a tibble), or a lazy data frame
(e.g. from dbplyr or dtplyr).}

\item{.keep_all}{If TRUE, keep all variables in .data. If a combination of expr is not distinct,
this keeps the first row of values.}
}
\description{
This function is similar to R function \code{dplyr::distinct}.
}
