% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.check_data}
\alias{.check_data}
\title{Check Data}
\usage{
.check_data(.data)
}
\arguments{
\item{.data}{The data object to be checked.}
}
\description{
This function performs checks on the specified data object, including existence and class.
}
