% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.check_exceeds_threshold}
\alias{.check_exceeds_threshold}
\title{Check Which Variables Exceed a Threshold}
\usage{
.check_exceeds_threshold(variable_names, variable_lengths, threshold)
}
\arguments{
\item{variable_names}{A character vector containing the names of the variables.}

\item{variable_lengths}{A numeric vector containing the lengths of the variables, in the same order as `variable_names`.}

\item{threshold}{A numeric value specifying the threshold that each variable length is compared against.}
}
\value{
A character vector of variable names where the corresponding lengths exceed the threshold.
}
\description{
This function checks if the lengths of the given variables exceed a specified threshold.
It returns the names of the variables whose lengths are greater than the threshold.
}
