% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.set_privacy_option}
\alias{.set_privacy_option}
\title{Set Privacy Option
Retrieves the value of a given privacy setting or set a default}
\usage{
.set_privacy_option(setting)
}
\arguments{
\item{setting}{The name of the privacy setting.}
}
\value{
The value of the specified privacy setting.
}
\description{
Set Privacy Option
Retrieves the value of a given privacy setting or set a default
}
\keyword{internal}
