% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dsListDisclosureSettingsTidyVerse}
\alias{dsListDisclosureSettingsTidyVerse}
\title{List Disclosure Settings
Outputs a list of disclosure settings}
\usage{
dsListDisclosureSettingsTidyVerse()
}
\value{
A named list of disclosure settings.
}
\description{
List Disclosure Settings
Outputs a list of disclosure settings
}
