% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupKeysDS.R
\name{groupKeysDS}
\alias{groupKeysDS}
\title{Performs dplyr \code{group_keys}.}
\usage{
groupKeysDS(x)
}
\arguments{
\item{x}{a grouped tibble.}

\item{tidy_select}{Unused in this function.}
}
\value{
A tibble describing the groups is returned to the client.
}
\description{
Performs dplyr \code{group_keys}.
}
