% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ifElseDS.R
\name{ifElseDS}
\alias{ifElseDS}
\title{Vectorised if-else}
\usage{
ifElseDS(
  condition = NULL,
  true = NULL,
  false = NULL,
  missing = NULL,
  ptype = NULL,
  size = NULL
)
}
\arguments{
\item{condition}{A list, specifying a logical vector in tidyverse syntax, ie data and column names unquoted.}

\item{true}{Vector to use for TRUE value of condition.}

\item{false}{Vector to use for FALSE value of condition.}

\item{missing}{If not NULL, will be used as the value for NA values of condition. Follows the same size and type rules as true and false.}

\item{ptype}{An optional prototype declaring the desired output type. If supplied, this overrides the common type of true, false, and missing.}

\item{size}{An optional size declaring the desired output size. If supplied, this overrides the size of condition.}
}
\value{
the object specified by the \code{newobj} argument of \code{ds.if_else} which is written to the serverside.
}
\description{
DataSHIELD implentation of \code{dply::if_else}.
}
