% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{listPermittedTidyverseFunctionsDS}
\alias{listPermittedTidyverseFunctionsDS}
\title{List of Permitted Tidyverse Functions}
\usage{
listPermittedTidyverseFunctionsDS()
}
\value{
A character vector of function names, each representing a permitted function. Functions
not included in this list will be blocked.
}
\description{
This function returns a vector of function names that are permitted to be passed within the
dsTidyverse functions, e.g. within the `tidy_select` argument of `ds.mutate.`
}
