% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renameDS.R
\name{renameDS}
\alias{renameDS}
\title{Rename columns}
\usage{
renameDS(tidy_expr, .data)
}
\arguments{
\item{tidy_expr}{Diffused expression passed to ds.rename.}

\item{.data}{A data frame or tibble.}
}
\value{
the object specified by the \code{newobj} argument of \code{ds.rename} which is written
to the serverside.
}
\description{
DataSHIELD implentation of\code{dplyr::rename}.
}
