% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectDS.R
\name{selectDS}
\alias{selectDS}
\title{Keep or drop columns using their names and types}
\usage{
selectDS(tidy_expr, .data)
}
\arguments{
\item{tidy_expr}{One or more unquoted expressions separated by commas.}

\item{.data}{A data frame or tibble.}
}
\value{
the object specified by the \code{newobj} argument of \code{ds.select} which is written
to the serverside.
}
\description{
DataSHIELD implentation of \code{dplyr::select}.
}
\details{
Performs dplyr select
}
