% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupByDS.R
\name{ungroupDS}
\alias{ungroupDS}
\title{Remove grouping from a tibble or data frame}
\usage{
ungroupDS(tidy_expr, x)
}
\arguments{
\item{tidy_expr}{Unused in this function.}

\item{x}{A tibble.}
}
\value{
An ungrouped data frame or tibble.
}
\description{
DataSHIELD implentation of \code{dplyr::ungroup}.
}
