% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrangeDS.R
\name{arrangeDS}
\alias{arrangeDS}
\title{Order the rows of a data frame by the values of selected columns}
\usage{
arrangeDS(tidy_expr, df.name, .by_group)
}
\arguments{
\item{tidy_expr}{Variables, or functions of variables. Use \code{desc()} to sort a variable in descending
order.}

\item{df.name}{A data frame, data frame extension (e.g. a tibble), or a lazy data frame (e.g. from dbplyr or dtplyr).}

\item{.by_group}{If TRUE, will sort first by grouping variable. Applies to grouped data frames
only.}
}
\value{
An object of the same type as \code{df.name}, typically a data frame or tibble.
}
\description{
DataSHIELD implentation of \code{dplyr::arrange}.
}
