% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupBy.R
\name{groupByDS}
\alias{groupByDS}
\title{Performs dplyr filter}
\usage{
groupByDS(expr, .data, .add, .drop)
}
\arguments{
\item{expr}{Diffused expression of dotdotdot passed to ds.filter}

\item{.data}{A data frame, data frame extension (e.g. a tibble), or a lazy data frame
(e.g. from dbplyr or dtplyr).}

\item{.add}{When FALSE, the default, \code{group_by()} will override existing groups. To add to
the existing groups, use .add = TRUE.}

\item{.drop}{Drop groups formed by factor levels that don't appear in the data? The default is
TRUE except when .data has been previously grouped with .drop = FALSE.}
}
\value{
the object specified by the \code{newobj} argument of \code{ds.group_by} which is written
to the serverside.
}
\description{
This function is similar to R function \code{filter}.
}
\details{
Performs dplyr filter
}
