% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupKeysDS.R
\name{groupKeysDS}
\alias{groupKeysDS}
\title{Performs dplyr \code{group_keys}.}
\usage{
groupKeysDS(tidy_select, x)
}
\arguments{
\item{tidy_select}{Unused in this function.}

\item{x}{a grouped tibble.}
}
\value{
A tibble describing the groups is returned to the client.
}
\description{
Performs dplyr \code{group_keys}.
}
