% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renameDS.R
\name{renameDS}
\alias{renameDS}
\title{Rename columns}
\usage{
renameDS(tidy_expr, df.name)
}
\arguments{
\item{tidy_expr}{list containing diffused expression.}

\item{df.name}{A data frame or tibble.}
}
\value{
the object specified by the \code{newobj} argument of \code{ds.rename} which is written
to the serverside.
}
\description{
DataSHIELD implentation of\code{dplyr::rename}.
}
