% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renameDS.R
\name{renameDS}
\alias{renameDS}
\title{Performs dplyr rename}
\usage{
renameDS(.data, expr)
}
\arguments{
\item{.data}{A data frame, data frame extension (e.g. a tibble), or a lazy data frame (e.g. from dbplyr or dtplyr).}

\item{expr}{Diffused expression of dotdotdot passed to ds.rename}
}
\value{
the object specified by the \code{newobj} argument
of \code{ds.rename} (or as default same name as input object) which is written to the serverside.
}
\description{
This function is similar to R function \code{rename}.
}
\details{
Performs dplyr rename
}
