% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectDS.R
\name{selectDS}
\alias{selectDS}
\title{Keep or drop columns using their names and types}
\usage{
selectDS(tidy_expr, df.name)
}
\arguments{
\item{tidy_expr}{One or more unquoted expressions separated by commas.}

\item{df.name}{A data frame or tibble.}
}
\value{
An object of the same type as \code{df.name}, typically a data frame or tibble.
}
\description{
DataSHIELD implentation of \code{dplyr::select}.
}
\details{
Performs dplyr select
}
