% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectDS.R
\name{selectDS}
\alias{selectDS}
\title{Performs dplyr select}
\usage{
selectDS(expr, .data)
}
\arguments{
\item{expr}{Diffused expression of dotdotdot passed to ds.select}

\item{.data}{A data frame, data frame extension (e.g. a tibble), or a lazy data frame (e.g. from dbplyr or dtplyr).}
}
\value{
the object specified by the \code{newobj} argument of \code{ds.select} which is written
to the serverside.
}
\description{
This function is similar to R function \code{select}.
}
\details{
Performs dplyr select
}
