% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupByDS.R
\name{ungroupDS}
\alias{ungroupDS}
\title{Performs dplyr ungroup}
\usage{
ungroupDS(tidy_select, x)
}
\arguments{
\item{tidy_select}{Unused in this function.}

\item{x}{A tibble.}
}
\value{
the object specified by the \code{newobj} argument of \code{ds.ungroup} which is written
to the serverside.
}
\description{
This function is similar to R function \code{dplyr::ungroup}.
}
